ct: Add additional system tests.
- This version of eslint doesn't support eslint.json.
- Add a script for running tests in debug mode (Node 6.8.0 only).
- Add a few additional system tests, including one that does an update
  w/o a merge function (to verify the fix for
[PR2](https://github.com/IBM/node-nano-doc-updater/pull/2)).
