+++ .gitignore
@@ -1,2 +1,3 @@
 /.project
 node_modules
+npm-debug.log
