+++ test/system/test-reset-test-db.js
@@ -1,0 +1,31 @@
+var test = require("tape-catch"),
+    Promise = require("bluebird"),
+    nanoDocUpdater = require("../../"),
+    resetTestDb = require("./lib/reset-test-db");
+
+test("sanity - after wiping the DB...", (t) => {
+    resetTestDb()
+    .then((db) => {
+        db = Promise.promisifyAll(db);
+        return db.listAsync();
+    })
+    .then((r) => {
+        var body = r[0];
+
+        t.deepEquals(body.rows, [], "...is the db empty?");
+    })
+    .catch((e) => {
+        if (e instanceof Error) {
+            t.comment(e.stack);
+            t.fail("...an error occurred");
+        } else {
+            t.comment(e.stack);
+            t.fail("...an error occurred");
+        }
+
+        throw(e);
+    })
+    .finally(() => {
+        t.end();
+    });
+});
